requirements.txt
asgiref==3.11.0
boto3==1.42.52
botocore==1.42.52
Django==5.2.8
django-environ==0.12.0
django-stdimage==6.0.2
django-storages==1.14.6
gunicorn==25.1.0
jmespath==1.1.0
packaging==26.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
